# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DriveAssist API",
        "description": "API documentation for DriveAssist - Vehicle Diagnostics and Expert Marketplace",
        "contact": {
            "name": "DriveAssist Support",
            "email": "support@driveassist.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/diagnoses": {
            "get": {
                "tags": [
                    "Diagnoses"
                ],
                "summary": "List user diagnoses",
                "operationId": "ab672c52f3f59154c668ba6389d926ef",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Diagnoses"
                ],
                "summary": "Create a new diagnosis",
                "operationId": "2e9df781414b28078eec5a25b23a9342",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "vehicle_id": {
                                        "type": "integer"
                                    },
                                    "symptoms_description": {
                                        "type": "string"
                                    },
                                    "region_id": {
                                        "type": "integer"
                                    },
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Diagnosis created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/diagnoses/guest": {
            "post": {
                "tags": [
                    "Diagnoses"
                ],
                "summary": "Create a guest diagnosis",
                "operationId": "e72318c4054aba594e4b8dff5f03b21d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "symptoms_description"
                                ],
                                "properties": {
                                    "symptoms_description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Diagnosis created"
                    }
                }
            }
        },
        "/api/v1/diagnoses/{id}": {
            "get": {
                "tags": [
                    "Diagnoses"
                ],
                "summary": "Get diagnosis details",
                "operationId": "5c4a81f016ecb7b2193369746cf96b4f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}